+++ src/PRD/Instance.h
@@ -41,7 +41,7 @@ public :
 	/// <para>indice 1 -> dépot distribution</para>
 	/// <para>les autres indices -> lieu de livraison des jobs (2 -> job 0, 3 -> job 1 ,...)</para>
 	/// </summary>
-	vector<vector<int>> t;
+	vector<vector<double>> t;
 
 	/// <summary>
 	/// Très grande valeur arbitraire
